Looks improvement.

+++ views/questions/index.php
@@ -5,14 +5,29 @@
         <h3><a href="/questions/create">Ask a question</a></h3>
     <?php endif ?>
 
-    <ul class="questionsList">
+    <ul class="tagsSidebar">
+        <li><strong>Tags: </strong></li>
+        <li><a href="/questions">All</a></li>
+        <?php foreach ($this->tags as $tag) : ?>
+            <li>
+                <a href="questions/viewTag/<?= htmlspecialchars($tag['name']) ?> "><?= htmlspecialchars($tag['name']) ?></a>
+                <?php if (isset($_SESSION['isAdmin']) && $_SESSION['isAdmin'] > 0) : ?>
+                    <a href="questions/deleteTag/<?= htmlspecialchars($tag['id']) ?> " class="warningText">[X]</a>
+                <?php endif ?>
+            </li>
+        <?php endforeach ?>
+    </ul>
+
+
+    <div class="questionsList">
+    <ul>
         <?php foreach ($this->questions as $question) : ?>
             <li>
                 <a href="questions/viewQuestion/<?= $question['id'] ?> "><?= htmlspecialchars($question['title']) ?> [<?= htmlspecialchars($question['visits']) ?>]</a>
             </li>
         <?php endforeach ?>
     </ul>
-
+    </div>
 
 
     <ul class="categoriesSidebar">
@@ -21,7 +36,7 @@
         <?php foreach ($this->categories as $category) : ?>
             <li>
                 <a href="questions/viewCategory/<?= htmlspecialchars($category['name']) ?> "><?= htmlspecialchars($category['name']) ?></a>
-                <?php if ($_SESSION['isAdmin'] > 0) : ?>
+                <?php if (isset($_SESSION['isAdmin']) && $_SESSION['isAdmin'] > 0) : ?>
                     <a href="questions/deleteCategory/<?= htmlspecialchars($category['id']) ?> " class="warningText">[X]</a>
                 <?php endif ?>
             </li>
@@ -40,18 +55,7 @@
         </li>
     </ul>
 
-    <ul class="tagsSidebar">
-        <li><strong>Tags: </strong></li>
-        <li><a href="/questions">All</a></li>
-        <?php foreach ($this->tags as $tag) : ?>
-            <li>
-                <a href="questions/viewTag/<?= htmlspecialchars($tag['name']) ?> "><?= htmlspecialchars($tag['name']) ?></a>
-                <?php if ($_SESSION['isAdmin'] > 0) : ?>
-                    <a href="questions/deleteTag/<?= htmlspecialchars($tag['id']) ?> " class="warningText">[X]</a>
-                <?php endif ?>
-            </li>
-        <?php endforeach ?>
-    </ul>
+
 
 
 </div>
